requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.5.2
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
Flask==3.0.3
Flask-Dance==7.1.0
Flask-Login==0.6.3
fonttools==4.54.1
fsspec==2024.10.0
gradio==4.44.1
gradio_client==1.3.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
numpy==1.24.4
oauthlib==3.2.2
openai==1.54.4
orjson==3.10.11
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
ruff==0.7.3
safehttpx==0.1.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
tomlkit==0.12.0
tqdm==4.67.0
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
URLObject==2.4.3
uvicorn==0.32.0
watchdog==4.0.2
websockets==12.0
Werkzeug==3.0.6
zipp==3.20.2
flask-restful==0.3.10
flask-cors>=4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
PyMySQL==1.1.0
cryptography==42.0.5
tiktoken
flask-mail
flask_migrate
stripe
mysqlclient
psycopg2
pika
